Time: 88 ms (33.06%) | Memory: 148.6 MB (30.01%) - LeetSync

+++ 295-find-median-from-data-stream/find-median-from-data-stream.cpp
@@ -32,7 +32,7 @@ public:
          if(maxheap.size() == minheap.size()) {
             return (maxheap.top() + minheap.top()) / 2.0;
         }
-        return (double)maxheap.top();
+        return maxheap.top();
     }
 };
